More cleanup

+++ emergency_eval/eval_single_long.py
@@ -1,26 +1,44 @@
+import copy
 import os
-
 import pickle
 
-import numpy as np
-
-from emergency_eval.run_eval import compare_results, compare_results_single_stream
+from emergency_eval import settings
+from emergency_eval.run_eval import compare_results_single_stream
 from emergency_eval.settings import EVAL_PATH_SIM
-from eval import extract_data, load_eval_files, check_arrival_delays, calc_metrics
+from eval import extract_data, load_eval_files, calc_offset_to_expected, calc_metrics, check_arrival_times
 
 
-def eval_for_path_with_run(path, run_num, results, results_merged):
+def eval_for_path_with_run(path, run_num, results):
     extract_data(path, run_num)
     streams, emergency_streams, streams_meta = load_eval_files(path, f"{path}/stream_meta.json", run_num)
-    check_arrival_delays(streams, streams_meta, True)
+    calc_offset_to_expected(streams, streams_meta)
 
-    if path not in results_merged:
-        results_merged[path] = {
-            "streams": streams,
-            "emergency_streams": emergency_streams,
-            "streams_meta": streams_meta,
-        }
-    else:
+    metrics = calc_metrics(streams, streams_meta, False)
+
+    if path not in results:
+        results[path] = {}
+
+    results[path][run_num] = {
+        'streams': streams,
+        'emergency_streams': emergency_streams,
+        'streams_meta': streams_meta,
+        'metrics': metrics
+    }
+
+
+def merge_runs_for_path(results, path, results_merged):
+    for run_num in results[path]:
+        if path not in results_merged:
+            results_merged[path] = {
+                'streams': copy.deepcopy(results[path][run_num]['streams']),
+                'emergency_streams': copy.deepcopy(results[path][run_num]['emergency_streams']),
+                'streams_meta': copy.deepcopy(results[path][run_num]['streams_meta']),
+            }
+            continue
+
+        streams = results[path][run_num]['streams']
+        emergency_streams = results[path][run_num]['emergency_streams']
+
         for port, stream in streams.items():
             intended_len = len(stream['delay'][0])
             if len(stream['delay'][0]) != intended_len:
@@ -34,6 +52,9 @@ def eval_for_path_with_run(path, run_num, results, results_merged):
             results_merged[path]["streams"][port]['delay'][0] += stream['delay'][0]
             results_merged[path]["streams"][port]['delay'][1] += stream['delay'][1]
             results_merged[path]["streams"][port]['offset_to_expected'] += stream['offset_to_expected']
+            results_merged[path]["streams"][port]['too_early'] += stream['too_early']
+            results_merged[path]["streams"][port]['too_late'] += stream['too_late']
+            results_merged[path]["streams"][port]['delayed'] += stream['delayed']
 
         for port, emergency_stream in emergency_streams.items():
             if port not in results_merged[path]["emergency_streams"]:
@@ -42,17 +63,8 @@ def eval_for_path_with_run(path, run_num, results, results_merged):
                 results_merged[path]["emergency_streams"][port]['delay'][0] += emergency_stream['delay'][0]
                 results_merged[path]["emergency_streams"][port]['delay'][1] += emergency_stream['delay'][1]
 
-    metrics = calc_metrics(streams, streams_meta, False)
-
-    if path not in results:
-        results[path] = {}
-
-    results[path][run_num] = {
-        'streams': streams,
-        'emergency_streams': emergency_streams,
-        'streams_meta': streams_meta,
-        'metrics': metrics
-    }
+    metrics = calc_metrics(results_merged[path]["streams"], results_merged[path]["streams_meta"], True)
+    results_merged[path]["metrics"] = metrics
 
 
 def single_streams(results_merged):
@@ -75,7 +87,7 @@ def single_streams(results_merged):
             }, ignore_index=True)
             pass
         for emergency_stream_name, emergency_stream_data in folder_data["emergency_streams"].items():
-          pass
+            pass
         pass
 
 
@@ -90,41 +102,48 @@ if __name__ == "__main__":
 
     eval_folders = [
         et_out,
-        # et_out_2,
         lib_out
     ]
 
+    # Load results per run
     result_file = f"{scenario_folder}/results.pkl"
-    result_file_merged = f"{scenario_folder}/results_merged.pkl"
-
     results = {}
+
+    result_file_merged = f"{scenario_folder}/results_merged.pkl"
     results_merged = {}
 
-    if not os.path.exists(result_file):
-        print("Should not be here!")
-        raise Exception("Should not be here!")
-        for i in range(100):
-            for folder in eval_folders:
-                eval_for_path_with_run(folder, i, results, results_merged)
+    if not os.path.exists(result_file_merged):
+        if not os.path.exists(result_file):
+            print("Extracting data...")
+            for i in range(settings.num_runs):
+                for folder in eval_folders:
+                    eval_for_path_with_run(folder, i, results)
 
-        for folder in eval_folders:
-            metrics = calc_metrics(results_merged[folder]["streams"], results_merged[folder]["streams_meta"], True)
-            results_merged[folder]["metrics"] = metrics
+            with open(result_file, "wb") as f:
+                pickle.dump(results, f)
+        else:
+            print("Load results.pkl")
+            with open(result_file, "rb") as f:
+                results = pickle.load(f)
 
-        # Save results as pickle
-        with open(result_file, "wb") as f:
-            pickle.dump(results, f)
+        # Merge results of multiple runs of the same folder
+        print("Merging results")
+        for path in results:
+            merge_runs_for_path(results, path, results_merged)
+
         with open(result_file_merged, "wb") as f:
             pickle.dump(results_merged, f)
     else:
-        # with open(result_file, "rb") as f:
-            # results = pickle.load(f)
+        print("Load results_merged.pkl")
         with open(result_file_merged, "rb") as f:
             results_merged = pickle.load(f)
 
+    # Perform evaluation on merged results
+    print("Sanity check")
     for folder in eval_folders:
+        check_arrival_times(results_merged[folder]["streams"])
+
         print(folder, results_merged[folder]["metrics"])
-        # Sanity check
         for port, stream in results_merged[folder]["streams"].items():
             intended_len = len(stream['delay'][0])
             if len(stream['delay'][0]) != intended_len:
@@ -134,12 +153,5 @@ if __name__ == "__main__":
             if len(stream['offset_to_expected']) != intended_len:
                 print(f"Error: {folder} {port} stream['offset_to_expected'][0] {len(stream['offset_to_expected'][0])} != {intended_len}")
 
-    # single_streams(results_merged)
-
-    # compare_results(results_merged,
-    #                 group_by_cycle=True)
-
+    print("Calc results")
     compare_results_single_stream(results_merged)
-
-
-
